cleaned up sync code.  Removed confusing ruby config files (only in sync_ipm.env now).  Use REXML to build osm document (much nicer AND should handle xml escaping properly)

+++ sync_ipm.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+# Script to synchronize formhub data with an osm instance
 
 # load config
 . sync_ipm.env
@@ -13,31 +14,34 @@ cp load_$OSM_ENV/existing.json load_$OSM_ENV/existing_bak.json
 curl -u $FH_USER:$FH_PWD https://formhub.org/$FH_USER/forms/$FH_FORM/api > load_$OSM_ENV/latest.json 
 
 # - Get diff of latest.json with existing.json and put it in new.json
-ruby fh_json_diff.rb
+if ! ./fh_json_diff.rb load_$OSM_ENV/existing.json load_$OSM_ENV/latest.json > load_$OSM_ENV/new.json; then
+    exit 1
+fi
 
 # open a changeset and get its id
 http_code=`curl -s -o changeset_id -w "%{http_code}" -u $OSM_USER:$OSM_PWD -d @changeset_new.xml -H "X_HTTP_METHOD_OVERRIDE: PUT" $OSM_SERVER/api/0.6/changeset/create`
 if [[ $http_code != 200 ]]; then
-    echo "FAILED to open changeset"
+    echo "FAILED to open changeset" >&2
     exit $http_code
 fi
 
-# store the changeset id in the cfg
+# load the changeset_id
 CHANGESET_ID=`cat changeset_id`
-sed "s/%change_id%/$CHANGESET_ID/" to_osm_cfg_tmpl.rb > to_osm_cfg.rb
 
 # create the upload osm changeset file from new.json
-cat load_$OSM_ENV/new.json | ruby json_to_osm.rb > load_$OSM_ENV/osm_upload.osc
+if ! cat load_$OSM_ENV/new.json | ./json_to_osm.rb $CHANGESET_ID change > load_$OSM_ENV/osm_upload.osc; then
+    exit 1
+fi
 
 # upload the changeset
 http_code=`curl -s -o diff_response -w "%{http_code}" -u $OSM_USER:$OSM_PWD -d @osm_upload.osc $OSM_SERVER/api/0.6/changeset/$CHANGESET_ID/upload`
 if [[ $http_code != 200 ]]; then
-    echo "FAILED to upload data"
+    echo "FAILED to upload data" >&2
 else
     # SUCCESS, so merge latest with existing dataset
     # ** canNOT just overwrite with latest since we may be pulling
     # ** from multiple forms
-    ruby -e "require 'json'; existing = JSON.parse(File.read(\"load_$OSM_ENV/existing.json\")); new = JSON.parse(File.read(\"load_$OSM_ENV/new.json\")); merged = existing + new;  puts merged.to_json;" > load_$OSM_ENV/merged.json
+    ruby -e "require 'json'; existing = JSON.parse(File.read(\"load_$OSM_ENV/existing.json\")); new_json = JSON.parse(File.read(\"load_$OSM_ENV/new.json\")); merged = existing + new_json;  puts merged.to_json;" > load_$OSM_ENV/merged.json
     # Is this necessary?  Or can we write to existing.json in cmd above
     mv load_$OSM_ENV/merged.json load_$OSM_ENV/existing.json
 fi
